Support $handle arrays in style enqueueing
`wp_enqueue_style()`, `wp_dequeue_style()`, and `wp_deregister_style()` already accept an array of strings for the `$handle` parameter. This behavior is not documented in the docblocks: the docblocks incorrectly suggest that only strings are allowed.

For `wp_enqueue_style()`,  in order to formally support arrays, I added a variable type check to make sure that `$handle` is a string before attempting to `explode()` it.